Add CI config
You'll notice that this doesn't actually run any tests because we don't
have any here. This is more of a build CI at this point, which is useful
in that it installs dependencies and compiles the JS & CSS assets in a
new environment.

This should help prevent breaking the app for interviewees.